extract ClubTabCard component out of ClubTabCards
